Move panel related objects to a dedicated module file to improve readability
