+++ packages/application/src/index.ts
@@ -3,3 +3,4 @@
 
 export * from './app';
 export * from './shell';
+export * from './panelhandler';
